split transactions into send and recv structs
